+++ backend/Dbo/PizzasIngredient.cs
@@ -2,8 +2,8 @@
 {
     public class PizzasIngredient
     {
-        public long IngredientId { get; set; }
-        public long PizzaId { get; set; }
+        public int IngredientId { get; set; }
+        public int PizzaId { get; set; }
 
         public virtual Ingredient Ingredient { get; set; }
         public virtual Pizza Pizza { get; set; }
